+++ CHANGELOG.md
@@ -4,6 +4,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project tries to adhere to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.7.0] - 2020-04-09
+### Added 
+- real station data
+- icons and images to landing page
+
+### Changed
+- timed display route to carousel
+- Departure station can be changed
+- category icons to filled and line icons
+- Answer animation can be skipped
+
+### Fixed
+- First dashboard popup layout
+
 ## [0.6.0] - 2020-02-19
 ### Added
 - Environment page.